refactor: adjust whole program structure to enhance modularity (SoC)
